JSON number is float64, convert to int64.
I'm setting the planned_end_date to an int64 in timelord, but it's still
getting parsed out as a float64 here. Since I was converting the
interface{} to an int64, things were blowing up. The code down the line
needs an int64, so convert it immediately to hopefully prevent further
panics.